+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 # [Unreleased](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v19.0.1...HEAD)
-* [STREAM-357](https://inindca.atlassian.net/browse/STREAM-357) - Maintain the same JID across hard reconnects/websockets/stanza instances so that reinvites can be sent to the new WS/stanza instance.
+* [STREAM-357](https://inindca.atlassian.net/browse/STREAM-357) - Maintain the same JID resource across hard reconnects/websockets/stanza instances so that reinvites can be sent to the new WS/stanza instance.
 
 # [v19.0.1](https://github.com/purecloudlabs/genesys-cloud-streaming-client/compare/v19.0.0...v19.0.1)
 ### Fixed